+++ app/controllers/datasets_controller.rb
@@ -2,9 +2,9 @@ class DatasetsController < ApplicationController
   before_action :authenticate_user_from_token!, only: [ :manifest, :files ]
   before_action :authenticate_user!,        only: [ :new, :edit, :create, :update, :destroy, :audits, :requests, :request_access, :set_access, :create_access, :new_page, :create_page, :edit_page, :update_page, :pull_changes, :sync ]
   before_action :check_system_admin,        only: [ :new, :create, :destroy, :pull_changes, :sync ]
-  before_action :set_viewable_dataset,      only: [ :show, :manifest, :logo, :images, :variable_chart, :files, :pages, :request_access, :search ]
+  before_action :set_viewable_dataset,      only: [ :show, :manifest, :logo, :images, :files, :pages, :request_access, :search ]
   before_action :set_editable_dataset,      only: [ :edit, :update, :destroy, :audits, :requests, :set_access, :create_access, :new_page, :create_page, :edit_page, :update_page, :pull_changes, :sync ]
-  before_action :redirect_without_dataset,  only: [ :show, :manifest, :logo, :images, :variable_chart, :files, :pages, :request_access, :set_access, :create_access, :search, :edit, :update, :destroy, :audits, :requests, :new_page, :create_page, :edit_page, :update_page, :pull_changes, :sync ]
+  before_action :redirect_without_dataset,  only: [ :show, :manifest, :logo, :images, :files, :pages, :request_access, :set_access, :create_access, :search, :edit, :update, :destroy, :audits, :requests, :new_page, :create_page, :edit_page, :update_page, :pull_changes, :sync ]
 
   # Concerns
   include Pageable
@@ -65,18 +65,6 @@ class DatasetsController < ApplicationController
 
   def logo
     send_file File.join( CarrierWave::Uploader::Base.root, @dataset.logo.url )
-  end
-
-  def variable_chart
-    name = params[:name].to_s.gsub(/[^\w\d-]/, '')
-
-    chart_file = File.join( @dataset.root_folder, 'dd', 'graphs', "#{name}.png")
-
-    if File.file?(chart_file)
-      send_file chart_file
-    else
-      render nothing: true
-    end
   end
 
   def files